+++ +context/UserContext.js
@@ -1,0 +1,6 @@
+import React from "react";
+
+export const UserContext = React.createContext({
+  user: undefined,
+  setUser: () => {}
+});
