+++ screens/MainScreen.js
@@ -1,6 +1,36 @@
 import React from "react";
 import { RegistrationScreen } from "./RegistrationScreen";
 
-export default function MainScreen() {
+import { Platform, StatusBar, StyleSheet, View } from "react-native";
+import { NavigationContainer } from "@react-navigation/native";
+import { createStackNavigator } from "@react-navigation/stack";
+
+import BottomTabNavigator from "../navigation/BottomTabNavigator";
+
+const Stack = createStackNavigator();
+
+export const MainScreen = ({ user, containerRef, initialNavigationState }) => {
+  if (user) {
+    return (
+      <View style={styles.container}>
+        {Platform.OS === "ios" && <StatusBar barStyle="default" />}
+        <NavigationContainer
+          ref={containerRef}
+          initialState={initialNavigationState}
+        >
+          <Stack.Navigator>
+            <Stack.Screen name="Root" component={BottomTabNavigator} />
+          </Stack.Navigator>
+        </NavigationContainer>
+      </View>
+    );
+  }
   return <RegistrationScreen />;
-}
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff"
+  }
+});
